requirements.txt
# Core dependencies for the medical document processing scripts
PyMuPDF==1.23.8
pytesseract==0.3.10
PyPDF2==3.0.1
google-genai==0.3.2
Pillow==10.1.0
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2

# Streamlit and web app dependencies
streamlit==1.28.1
watchdog==3.0.0

# Additional dependencies that might be needed
python-dotenv==1.0.0
